break position helper into seperate utils for easier unit testing
