+++ projects/swimlane/ngx-ui/src/lib/utils/position/calculate-horizontal-alignment/index.ts
@@ -1,0 +1,1 @@
+export * from './calculate-horizontal-alignment.util';
