+++ projects/swimlane/ngx-ui/src/lib/utils/position/calculate-horizontal-caret/index.ts
@@ -1,0 +1,1 @@
+export * from './calculate-horizontal-caret.util';
