+++ projects/swimlane/ngx-ui/src/lib/utils/position/calculate-vertical-alignment/index.ts
@@ -1,0 +1,1 @@
+export * from './calculate-vertical-alignment.util';
