+++ projects/swimlane/ngx-ui/src/lib/utils/position/calculate-vertical-caret/index.ts
@@ -1,0 +1,1 @@
+export * from './calculate-vertical-caret.util';
