+++ projects/swimlane/ngx-ui/src/lib/utils/position/caret-offset.constant.ts
@@ -1,0 +1,1 @@
+export const CARET_OFFSET = 7;
