+++ projects/swimlane/ngx-ui/src/lib/utils/position/determine-placement/determine-placement.util.ts
@@ -1,0 +1,33 @@
+import { PlacementTypes } from '../placement-type.enum';
+import { Dimensions } from '../dimensions.interface';
+import { AlignmentTypes } from '../alignment-types.enum';
+import { shouldFlip } from '../should-flip';
+
+/**
+ * Determine placement based on flip
+ *
+ * @param placement
+ * @param elmDim
+ * @param hostDim
+ * @param spacing
+ * @param alignment
+ *
+ * @memberOf PositionHelper
+ */
+export function determinePlacement(placement: PlacementTypes, elmDim: Dimensions, hostDim: Dimensions, spacing: number, alignment: AlignmentTypes) {
+  const flip = shouldFlip(hostDim, elmDim, placement, alignment, spacing);
+
+  if (flip) {
+    if (placement === PlacementTypes.right) {
+      return PlacementTypes.left;
+    } else if (placement === PlacementTypes.left) {
+      return PlacementTypes.right;
+    } else if (placement === PlacementTypes.top) {
+      return PlacementTypes.bottom;
+    } else if (placement === PlacementTypes.bottom) {
+      return PlacementTypes.top;
+    }
+  }
+
+  return placement;
+}
