+++ projects/swimlane/ngx-ui/src/lib/utils/position/horizontal-position/index.ts
@@ -1,0 +1,1 @@
+export * from './horizontal-position.util';
