+++ projects/swimlane/ngx-ui/src/lib/utils/position/index.ts
@@ -1,4 +1,16 @@
 export * from './placement-type.enum';
-export * from './position';
 export * from './dimensions.interface';
 export * from './alignment-types.enum';
+export * from './caret-offset.constant';
+
+// utils
+export * from './calculate-horizontal-alignment';
+export * from './calculate-horizontal-caret';
+export * from './calculate-vertical-alignment';
+export * from './calculate-vertical-caret';
+export * from './determine-placement';
+export * from './horizontal-position';
+export * from './position-caret';
+export * from './position-content';
+export * from './should-flip';
+export * from './vertical-position';
