+++ projects/swimlane/ngx-ui/src/lib/utils/position/position-content/index.ts
@@ -1,0 +1,1 @@
+export * from './position-content.util';
