+++ projects/swimlane/ngx-ui/src/lib/utils/position/should-flip/index.ts
@@ -1,0 +1,1 @@
+export * from './should-flip.util';
