+++ projects/swimlane/ngx-ui/src/lib/utils/position/vertical-position/index.ts
@@ -1,0 +1,1 @@
+export * from './vertical-position.util';
